Attempted to handle unusual passions better
For mods that somehow add in extra values to the passion enum (pretty sure this is impossible), or otherwise manage to return extra passion types from SkillRecord.passion, this will try to work out a reasonable cut-off based on the learning factor that SkillRecord then reports.

This also breaks if the fast learning debug option is turned on, as that forces learn rate to be 20000% no matter what.